Add launch-hidden functionality to moonshot msrpc implementation
Also, when the identity manager cannot bind its rpc interface,
it now binds as a client and tells the existing server to show
itself, similar to the dbus implementation.

To link both client and server rpc code requires passing the
'prefix' argument to midl which is not exposed via the mingw
midl wrapper, so I created new moonshot-midl wrapper to pass to the mingw
midl wrapper.