https://bugs.launchpad.net/moonshot-ui/+bug/1566560 (UI flashes on screen when moonshot is (re)started on command line)
When moonshot is launched on the command line, it was creating the UI window before it
even knows that the UI window is required, which is contingent on registering with DBus.
I've modified it to defer the creation of the window until it knows that it has
successfully registered with DBus.